% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-cont_electrode.R
\name{LFP_electrode}
\alias{LFP_electrode}
\title{Definition for electrodes that collects 'LFP' signals}
\description{
Definition for electrodes that collects 'LFP' signals

Definition for electrodes that collects 'LFP' signals
}
\examples{
\dontrun{

# Download demo subject DemoSubject ~700MB data

# Electrode 14
e <- LFP_electrode$new(subject = 'demo/DemoSubject',
                       number = 14, is_reference = FALSE)

# Reference "ref_13-16,24"
ref <- LFP_electrode$new(subject = 'demo/DemoSubject',
                         number = "ref_13-16,24", is_reference = TRUE)

# ------ Reference ------
# By default there is no reference
e$reference_name     # "noref

# set reference
e$set_reference(reference = ref)
e$reference_name     # "ref_13-16,24"

# Set epoch
e$set_epoch(epoch = 'auditory_onset')

# Now epoch power
power <- e$epoch_power(before_onset = 1, after_onset = 2)

# Trial x Frequency x Time x Electrodes
power

# Subset power
subset(power, Time ~ Time < 0, Electrode ~ Electrode == 14)

# clear memory in RAM and cache on hard disk
e$clear_cache()
e$clear_memory()


}
}
\section{Super classes}{
\code{\link[ravebase:RAVEAbstarctElectrode]{ravebase::RAVEAbstarctElectrode}} -> \code{\link[ravebase:Continuous_Electrode]{ravebase::Continuous_Electrode}} -> \code{LFP_electrode}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{type}}{type of signals collected by the electrode}
}
\if{html}{\out{</div>}}
}
