% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-repository.R
\name{loaded_rave_repository}
\alias{loaded_rave_repository}
\title{Create or get \code{\link{RAVERepository}} instance}
\usage{
loaded_rave_repository(subject, reference, epoch, before_onset, after_onset)
}
\arguments{
\item{subject}{character or \code{\link{RAVESubject}} instance}

\item{reference}{reference name}

\item{epoch}{epoch name}

\item{before_onset}{seconds to be loaded before trial onset}

\item{after_onset}{seconds to be loaded after trial onset}
}
\value{
\code{\link{RAVERepository}} instance
}
\description{
If repository has been loaded with the same signature, the
loaded one will be returned, otherwise generate a new repository
}
\examples{
\dontrun{

# Create repository for demo subject 'demo/DemoSubject',
# reference is reference_default.csv
# epoch is reference_auditory_onset.csv
# Time span is -1 ~ 2 (total 3 seconds, 0s is auditory onset)
repo <-
  loaded_rave_repository(
    subject = 'demo/DemoSubject',
    reference = 'default',
    epoch = 'auditory_onset',
    before_onset = 1,
    after_onset = 2
  )

# Epoch and get power data
power <- repo$epoch_power()

power

# get repository in other place instead of creating a new one
repo2 <-
  loaded_rave_repository(
    subject = 'demo/DemoSubject',
    reference = 'default',
    epoch = 'auditory_onset',
    before_onset = 1,
    after_onset = 2
  )

identical(repo2, repo)

# remove disk data
repo$clear_cache()

# release RAM
repo$clear_memory()

}

}
\seealso{
\code{\link{RAVERepository}}
}
