% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core-condition.R
\name{rave-states}
\alias{rave-states}
\alias{rave_is_running}
\title{Check States of Current RAVE instances or 'shiny' instances}
\usage{
rave_is_running()
}
\value{
Logical, whether rave has
context higher/equaling to \code{'rave_running'}.
}
\description{
Check States of Current RAVE instances or 'shiny' instances
}
