% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ext-threeBrain.R
\name{rave_brain2}
\alias{rave_brain2}
\title{Create 3D brain instance}
\usage{
rave_brain2(
  subject,
  surfaces = "pial",
  use_141 = TRUE,
  recache = FALSE,
  clean_before_cache = FALSE,
  compute_template = FALSE,
  usetemplateifmissing = FALSE
)
}
\arguments{
\item{subject}{character or \code{RAVESubject} instance}

\item{surfaces}{surface types, usually stored as \code{'*h.xxx'} in
'FreeSurfer' folder or 'SUMA' folder}

\item{use_141}{whether to use 'SUMA' standard 141 brain if exists}

\item{recache, clean_before_cache}{whether to re-calculate cache and
clean cache before redo cache}

\item{compute_template}{whether to compute for template brain, usually
use it when 141 brain is used}

\item{usetemplateifmissing}{if surface data is missing, should
template brain be used? this requires 'MNI305' coordinates}
}
\value{
A \code{rave-brain} instance, see more at 3D viewer function
\code{\link[threeBrain]{freesurfer_brain2}}
}
\description{
Create 3D brain instance
}
\seealso{
\code{\link[threeBrain]{freesurfer_brain2}}
}
