% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\name{rave_cache_path}
\alias{rave_cache_path}
\title{Temporary path where large cached data should be stored}
\usage{
rave_cache_path(temp = FALSE, check = FALSE)
}
\arguments{
\item{temp}{whether use temporary path}

\item{check}{whether to ensure the folder; if true, then the folder will be
created if missing}
}
\value{
Absolute path of cache path.
}
\description{
Temporary path where large cached data should be stored
}
