% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core-repo_loader.R
\docType{data}
\name{rave_load}
\alias{rave_load}
\alias{rave_attach}
\alias{rave_attach.RAVERepository}
\alias{rave_attach.character}
\alias{rave_attach.RAVESubject}
\alias{attached_repo}
\alias{rave_attached}
\title{Load or attach 'RAVE' subject}
\format{
An object of class \code{ravebase_rave_attached} (inherits from \code{ravebase_readonly}) of length 14.
}
\usage{
rave_load(
  subject,
  epoch,
  reference,
  before_onset,
  after_onset,
  electrodes,
  attach = FALSE
)

rave_attach(x, ...)

\method{rave_attach}{RAVERepository}(x, ...)

\method{rave_attach}{character}(x, epoch, reference, before_onset, after_onset, electrodes, ...)

\method{rave_attach}{RAVESubject}(x, epoch, reference, before_onset, after_onset, electrodes, ...)

attached_repo(test = FALSE)

rave_attached
}
\arguments{
\item{subject}{character, the format is \code{project/subject}, for example
\code{"demo/DemoSubject"}}

\item{epoch}{character, name of epoch, located in subject's 'meta' folder,
with file name \code{"epoch_<name>.csv"}}

\item{reference}{character, name of reference scheme, located in subject's
'meta' folder, with file name \code{"reference_<name>.csv"}}

\item{before_onset}{seconds before trial onset}

\item{after_onset}{seconds after trial onset}

\item{electrodes}{electrodes to load, if missing then default to all
possible electrodes}

\item{attach}{whether to attach loaded repository for analysis}

\item{x}{either \code{\link{RAVERepository}}, \code{\link{RAVESubject}}
instance or character, or \code{NULL}. For character, it's identical to
\code{subject}; for \code{NULL}, current attached repository will be
detached.}

\item{...}{passed to other methods}

\item{test}{whether return \code{NULL} instead of raising errors when
no subject is attached}
}
\value{
\code{rave_load} and \code{rave_attach} returns loaded repository,
an instance of \code{\link{RAVERepository}} class. \code{attached_repo}
returns currently loaded repository. If no repository loaded then it raises
errors.
}
\description{
Load or attach 'RAVE' a subject to current analysis context.
The goal is to allow analysis and visualization to be focused on one that
subject.
}
\details{
\code{rave_attached} is a list providing various of methods that
handles attached subjects. It also allows 'RAVE' modules designed for
single-subject analysis to get information of what's to be analyzed.

\code{epoch} and \code{reference} are located in subject's meta folder. The
folder path is \code{"<RAVE data_path>/project/subject/rave/meta/"}. Epoch
files start with \code{"epoch_"}, and its name is decided by the characters
following the prefix. For example, an epoch named "auditory" will have file
name \code{"epoch_auditory.csv"}. This rule applies to reference files,
except that the prefix is \code{"reference_"}

\code{before_onset} and \code{after_onset} affects the amount of data
to be loaded. For example, \code{before_onset=1} and \code{after_onset=2}
will result in total 3 seconds of data to be loaded: 1 second before trial
onset and 2 after onset. The reason to load data before onset is to leave
certain time range for baseline. If the trial lasts for 2 seconds, loading
3 seconds of data allows around 1 second to be counted into baseline.
The trial onset time for each trial is defined at column \code{"Time"} in
the epoch file.

If \code{attach=TRUE}, then \code{rave_load} is equivalent to
\code{rave_attach}.
}
\keyword{datasets}
