% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{RAVEEpoch}
\alias{RAVEPreprocessSettings}
\alias{RAVEProject}
\alias{RAVESubject}
\alias{as_rave_project}
\alias{as_rave_subject}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{raveio}{\code{\link[raveio]{RAVEEpoch}}, \code{\link[raveio]{RAVEPreprocessSettings}}, \code{\link[raveio]{RAVEProject}}, \code{\link[raveio]{RAVESubject}}, \code{\link[raveio]{as_rave_project}}, \code{\link[raveio]{as_rave_subject}}}
}}

